// apps/admin-web/src/.generated/api/model/adminEmailConfigResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 1.0.0
 */

export interface AdminEmailConfigResponseDto {
  smtpHost: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  smtpPort: number;
  smtpSecure: boolean;
  smtpUser: string;
  smtpPasswordConfigured: boolean;
  from: string;
}