// apps/admin-web/src/.generated/api/model/updateAdminEmailConfigRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateAdminEmailConfigRequestDto {
  smtpHost: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  smtpPort: number;
  smtpSecure: boolean;
  smtpUser: string;
  /** 비워두면 기존 비밀번호를 유지합니다. */
  smtpPassword?: string;
  from: string;
}